Create KBC game.py
I created a small replica of the game show called 'Kaun Banega Crorepati' (KBC) that I grew up watching, I used dictionaries over lists in this program as they seemed more viable to use. This is a simple, easy to understand code. 

I will be bringing further complexities to this game in the future.